requirements.txt
celery
redis
Django
django-celery-results
djangorestframework
djangorestframework-api-key
djangorestframework-simplejwt
psycopg2-binary

